mps2-an385: drop uint16_t cast in nros_freertos_create_task
xTaskCreate's third parameter is configSTACK_DEPTH_TYPE, which on
Cortex-M3 defaults to StackType_t = uint32_t (per the ARM_CM3 portmacro).
The C wrapper was casting our already-uint32_t stack_words to uint16_t
before the call, silently truncating any depth > 65535 words (> 256 KB)
to a 0-word stack with a wild SP.

Surfaced by Phase 13.K1 in the autoware_sentinel project: with the
default executor capacities (NROS_EXECUTOR_MAX_CBS=96 → ARENA_SIZE
≈ 346 KB), nros::Executor's inline arena gets copied ~3× by the
NRVO-defeated `Result<Executor, _>` return path, inflating the
app_task_entry stack frame to ~1 MB. The 64 KB / 128 KB FreeRTOS task
stack overflowed below SRAM base and the very first push faulted with
v7M INVSTATE UsageFault — appearing as "z_declare_publisher hangs"
that never actually reached zenoh code. Lifting this 256 KB ceiling
lets app_task_entry request the larger stack (e.g. 512 KB) it really
needs, and the comp-all topology boots cleanly through executor.spin().

+++ packages/boards/nros-board-mps2-an385-freertos/build.rs
@@ -519,7 +519,12 @@ int nros_freertos_create_task(
     void *arg,
     uint32_t priority)
 {
-    BaseType_t ret = xTaskCreate(entry, name, (uint16_t)stack_words, arg,
+    /* configSTACK_DEPTH_TYPE defaults to StackType_t (uint32_t on Cortex-M3
+     * via portmacro.h). The previous (uint16_t) cast silently truncated
+     * stack depths > 65535 words (>256 KB), leaving tasks with a 0-word
+     * stack and a wild SP. Drop the cast — xTaskCreate accepts the full
+     * uint32_t we already declared in this wrapper. */
+    BaseType_t ret = xTaskCreate(entry, name, stack_words, arg,
                                  (UBaseType_t)priority, NULL);
     return (ret == pdPASS) ? 0 : -1;
 }
